Fixed color selection in conversion from uint hex values to float (0-1)

+++ src/display/fake_surface.go
@@ -18,7 +18,7 @@ func (s *FakeSurface) MoveTo(x float64, y float64) {
 	s.commands = append(s.commands, SurfaceCommand{Name: "MoveTo", Args: args})
 }
 
-func (s *FakeSurface) SetRgba(r, g, b, a float64) {
+func (s *FakeSurface) SetRgba(r, g, b, a uint) {
 	args := []interface{}{r, g, b, a}
 	s.commands = append(s.commands, SurfaceCommand{Name: "SetRgba", Args: args})
 }
